Extract enabled assertions
