+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/introduction/IntroducingBaristaAssertions.java
@@ -7,10 +7,10 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertChecked;
-import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertDisabled;
+import static com.schibsted.spain.barista.assertion.BaristaEnabledAssertions.assertDisabled;
+import static com.schibsted.spain.barista.assertion.BaristaEnabledAssertions.assertEnabled;
 import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed;
 import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertDrawable;
-import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertEnabled;
 import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertHint;
 import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertNotDisplayed;
 import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertNotExist;
